+++ pocketbase_helpers_cli/lib/pocketbase_helpers_cli.dart
@@ -5,3 +5,4 @@ export 'src/field_mapper.dart';
 export 'src/field.dart';
 export 'src/utils.dart';
 export 'src/collection.dart';
+export 'src/schema_source.dart';